Add OU noise class
