+++ .gitignore
@@ -12,3 +12,5 @@ training/twmvt.py
 
 REPORT.md
 REPORT-VALGRAD.md
+
+build
